backported from "trunk" - import operation uses delta generator to send deltas instead of 'replacement' window. - import operatioh shouldn't create tmp files in imported directory, as it could be readonly. - delta generator vdelta optimized (slotstable object is reused). - GTM used both to expand date in Id and Date keywords. - export editor doesn't copy-translate tmp file when no translation is needed. - HTTP requests supports "long" contents length.

+++ javasvn/src/org/tmatesoft/svn/core/internal/wc/SVNFileUtil.java
@@ -257,9 +257,11 @@ public class SVNFileUtil {
             srcChannel = is.getChannel();
             os = new FileOutputStream(tmpDst);
             dstChannel = os.getChannel();
-            long count = srcChannel.size();
-            while (count > 0) {
-                count -= dstChannel.transferFrom(srcChannel, 0, srcChannel.size());
+            long copied = 0;
+            long totalSize = srcChannel.size();
+            while (copied < totalSize) {
+                long toCopy = Math.min(1024*1024*1024, totalSize - copied);
+                copied += dstChannel.transferFrom(srcChannel, copied, toCopy);
             }
         } catch (IOException e) {
             SVNErrorMessage err = SVNErrorMessage.create(SVNErrorCode.IO_ERROR, "Cannot copy file ''{0}'' to ''{1}''", new Object[] {src, dst});
@@ -494,6 +496,9 @@ public class SVNFileUtil {
     }
     
     public static void deleteFile(File file) {
+        if (file == null) {
+            return;
+        }
         if (!isWindows || file.isDirectory() || !file.exists()) {
             file.delete();
             return;
